@@ -48,65 +47,68 @@
     "commit": "git-cz"
   },
   "dependencies": {
-    "@microsoft/microsoft-graph-client": "1.4.0",
+    "@microsoft/microsoft-graph-client": "1.5.2",
     "@semantic-release/exec": "^3.3.2",
     "adal-node": "0.1.28",
-    "async": "2.6.1",
+    "async": "2.6.2",
     "async-promises": "0.2.1",
     "azure-arm-cognitiveservices": "2.4.1",
     "azure-arm-resource": "7.3.0",
     "azure-arm-search": "^1.3.0-preview",
     "azure-arm-sql": "5.6.0",
     "azure-arm-website": "5.7.0",
-    "bluebird": "^3.5.3",
+    "bluebird": "^3.5.4",
     "body-parser": "1.18.3",
-    "botbuilder": "^4.1.7",
-    "botbuilder-ai": "^4.2.0",
-    "botbuilder-azure": "^4.2.0",
-    "botbuilder-choices": "^4.0.0-preview1.2",
-    "botbuilder-dialogs": "^4.2.0",
-    "botbuilder-prompts": "^4.0.0-preview1.2",
-    "botlib": "^0.1.10",
+    "botbuilder": "4.3.4",
+    "botbuilder-ai": "4.3.4",
+    "botbuilder-azure": "4.3.4",
+    "botbuilder-choices": "4.0.0-preview1.2",
+    "botbuilder-dialogs": "4.3.4",
+    "botbuilder-prompts": "4.0.0-preview1.2",
+    "botlib": "0.1.24",
     "chai": "4.2.0",
     "child_process": "^1.0.2",
-    "chokidar": "2.0.4",
-    "cli-spinner": "^0.2.8",
-    "csv-parse": "4.3.1",
-    "dotenv-extended": "2.3.0",
+    "chokidar": "2.1.5",
+    "cli-spinner": "^0.2.10",
+    "csv-parse": "4.3.4",
+    "dotenv-extended": "2.4.0",
     "express": "4.16.4",
     "express-promise-router": "3.0.3",
     "fs-extra": "7.0.1",
-    "fs-walk": "0.0.2",
     "ip": "^1.1.5",
+    "js-beautify": "^1.9.1",
     "localize": "0.4.7",
-    "marked": "0.6.0",
-    "mocha": "5.2.0",
+    "marked": "0.6.2",
+    "mocha": "6.0.2",
     "mocha-typescript": "1.1.17",
     "ms": "2.1.1",
     "ms-rest-azure": "2.6.0",
+    "ms-rest-js": "^1.0.1",
     "nexmo": "2.4.1",
-    "ngrok": "^3.1.0",
-    "nyc": "^13.1.0",
-    "opn": "^5.4.0",
+    "ngrok": "3.1.1",
+    "nyc": "13.3.0",
+    "opn": "6.0.0",
     "pragmatismo-io-framework": "1.0.19",
-    "process-exists": "^3.1.0",
+    "process-exists": "3.1.0",
     "public-ip": "^3.0.0",
     "reflect-metadata": "0.1.13",
-    "request-promise-native": "1.0.5",
+    "request-promise": "4.2.4",
+    "request-promise-native": "1.0.7",
     "scanf": "^1.0.2",
-    "sequelize": "4.42.0",
-    "sequelize-typescript": "0.6.7",
-    "shx": "^0.3.2",
-    "simple-git": "^1.107.0",
+    "sequelize": "^5.2.12",
+    "sequelize-typescript": "0.6.9",
+    "shx": "0.3.2",
+    "simple-git": "1.110.0",
     "sqlite3": "4.0.6",
     "strict-password-generator": "^1.1.2",
-    "swagger-client": "3.8.22",
-    "tedious": "4.1.3",
-    "ts-node": "8.0.2",
+    "swagger-client": "3.8.25",
+    "tedious": "6.1.0",
+    "temperature-js": "^0.1.0",
+    "ts-node": "8.0.3",
     "typedoc": "0.14.2",
     "typedoc-plugin-external-module-name": "^2.0.0",
-    "typedoc-plugin-markdown": "^1.1.25",
-    "typescript": "3.3.1",
+    "typedoc-plugin-markdown": "^1.1.27",
+    "typescript": "3.4.1",
     "url-join": "4.0.0",
     "vbscript-to-typescript": "^1.0.8",
     "wait-until": "0.0.2",
