@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -34,42 +34,110 @@
 
 import { TurnContext } from 'botbuilder';
 import { WaterfallStepContext } from 'botbuilder-dialogs';
-import { GBMinInstance } from 'botlib';
-const WaitUntil = require('wait-until');
+import { GBLog, GBMinInstance } from 'botlib';
+import * as request from 'request-promise-native';
+import urlJoin = require('url-join');
+import { GBAdminService } from '../../admin.gbapp/services/GBAdminService';
+import { AzureDeployerService } from '../../azuredeployer.gbapp/services/AzureDeployerService';
+import { GBDeployer } from './GBDeployer';
 
 /**
  * @fileoverview General Bots server core.
  */
 
+/**
+ * BASIC system class for extra manipulation of bot behaviour.
+ */
+class SysClass {
+  public min: GBMinInstance;
+  private readonly deployer: GBDeployer;
+
+  constructor(min: GBMinInstance, deployer: GBDeployer) {
+    this.min = min;
+    this.deployer = deployer;
+  }
+
+  public async wait(seconds: number) {
+    // tslint:disable-next-line no-string-based-set-timeout
+    const timeout = async (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+    await timeout(seconds * 1000);
+  }
+
+  public generatePassword() {
+    return GBAdminService.getRndPassword();
+  }
+
+  public async createABotFarmUsing(
+    botId: string,
+    username: string,
+    password: string,
+    location: string,
+    nlpAuthoringKey: string,
+    appId: string,
+    appPassword: string,
+    subscriptionId: string
+  ) {
+    const service = new AzureDeployerService(this.deployer);
+    await service.deployToCloud(
+      botId,
+      username,
+      password,
+      location,
+      nlpAuthoringKey,
+      appId,
+      appPassword,
+      subscriptionId
+    );
+  }
+
+  /**
+   * Generic function to call any REST API.
+   */
+  public async sendEmail(to, subject, body) {
+    // tslint:disable-next-line:no-console
+    GBLog.info(`[E-mail]: to:${to}, subject: ${subject}, body: ${body}.`);
+  }
+
+  /**
+   * Generic function to call any REST API.
+   */
+  public async httpGet(url: string, qs) {
+
+    const options = {
+        uri: urlJoin(url , qs)
+    };
+
+    return request.get(options);
+  }
+
+}
+
+/**
+ * Base services of conversation to be called by BASIC.
+ */
 export class DialogClass {
+
   public min: GBMinInstance;
   public context: TurnContext;
   public step: WaterfallStepContext;
+  public internalSys: SysClass;
 
-  constructor(min: GBMinInstance) {
+  constructor(min: GBMinInstance, deployer: GBDeployer) {
     this.min = min;
+    this.internalSys = new SysClass(min, deployer);
+  }
+
+  public sys(): SysClass {
+    return this.internalSys;
   }
 
   public async hear(cb) {
-    let idCallback = Math.floor(Math.random() * 1000000000000);
+    const idCallback = crypto.getRandomValues(new Uint32Array(16))[0];
     this.min.cbMap[idCallback] = cb;
-    await this.step.beginDialog('/hear', { id: idCallback});
+    await this.step.beginDialog('/hear', { id: idCallback });
   }
 
   public async talk(text: string) {
     return await this.context.sendActivity(text);
   }
-
-  /**
-   * Generic function to call any REST API.
-   */
-  public sendEmail(to, subject, body) {
-    // tslint:disable-next-line:no-console
-    console.log(`[E-mail]: to:${to}, subject: ${subject}, body: ${body}.`);
-  }
-
-  /**
-   * Generic function to call any REST API.
-   */
-  public post(url: string, data) {}
 }
