@@ -2,7 +2,7 @@
 |                                               ( )_  _                       |
 |    _ _    _ __   _ _    __    ___ ___     _ _ | ,_)(_)  ___   ___     _     |
 |   ( '_`\ ( '__)/'_` ) /'_ `\/' _ ` _ `\ /'_` )| |  | |/',__)/' _ `\ /'_`\   |
-|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| ( ) |( (_) )  |
+|   | (_) )| |  ( (_| |( (_) || ( ) ( ) |( (_| || |_ | |\__, \| (˅) |( (_) )  |
 |   | ,__/'(_)  `\__,_)`\__  |(_) (_) (_)`\__,_)`\__)(_)(____/(_) (_)`\___/'  |
 |   | |                ( )_) |                                                |
 |   (_)                 \___/'                                                |
@@ -38,14 +38,22 @@
 
 import { BotAdapter } from 'botbuilder';
 import { WaterfallDialog } from 'botbuilder-dialogs';
-import { IGBDialog } from 'botlib';
-import { GBMinInstance } from 'botlib';
+import { GBLog, GBMinInstance, IGBDialog } from 'botlib';
 import { AzureText } from 'pragmatismo-io-framework';
 import { Messages } from '../strings';
 import { KBService } from './../services/KBService';
 
-const logger = require('../../../src/logger');
+/**
+ * Dialog arguments.
+ */
+export class AskDialogArgs {
+  public questionId: number;
+  public fromFaq: boolean;
+}
 
+/**
+ * Handle the ask loop on knowledge base data or delegate to other services.
+ */
 export class AskDialog extends IGBDialog {
   /**
    * Setup dialogs flows and define services call.
@@ -55,154 +63,135 @@ export class AskDialog extends IGBDialog {
    */
   public static setup(bot: BotAdapter, min: GBMinInstance) {
     const service = new KBService(min.core.sequelize);
+    min.dialogs.add(new WaterfallDialog('/answerEvent', AskDialog.getAnswerEventDialog(service, min)));
+    min.dialogs.add(new WaterfallDialog('/answer', AskDialog.getAnswerDialog(min, service)));
+    min.dialogs.add(new WaterfallDialog('/ask', AskDialog.getAskDialog(min)));
+  }
 
-    min.dialogs.add(
-      new WaterfallDialog('/answerEvent', [
-        async step => {
-          if (step.options && step.options['questionId']) {
-            const question = await service.getQuestionById(min.instance.instanceId, step.options['questionId']);
-            const answer = await service.getAnswerById(min.instance.instanceId, question.answerId);
+  private static getAskDialog(min: GBMinInstance) {
+    return [
+      async step => {
+        const locale = step.context.activity.locale;
+        const user = await min.userProfile.get(step.context, {});
+        user.isAsking = true;
+        if (!user.subjects) {
+          user.subjects = [];
+        }
+        let text;
+        // Three forms of asking.
+        if (step.options && step.options.firstTime) {
+          text = Messages[locale].ask_first_time;
+        } else if (step.options && step.options.isReturning) {
+          text = Messages[locale].anything_else;
+        } else if (user.subjects.length > 0) {
+          text = Messages[locale].which_question;
+        } else {
+          throw new Error('Invalid use of /ask');
+        }
+        if (text.length > 0) {
+          return await step.prompt('textPrompt', text);
+        }
 
-            // Sends the answer to all outputs, including projector.
-
-            await service.sendAnswer(min.conversationalService, step, answer);
-
-            await step.replaceDialog('/ask', { isReturning: true });
-          }
+        return await step.next();
+      },
+      async step => {
+        if (step.result) {
+          return await step.replaceDialog('/answer', { query: step.result });
+        } else {
           return await step.next();
         }
-      ])
-    );
+      }
+    ];
+  }
 
-    min.dialogs.add(
-      new WaterfallDialog('/answer', [
-        async step => {
-          const user = await min.userProfile.get(step.context, {});
-          let text = step.options['query'];
-          if (!text) {
-            throw new Error(`/answer being called with no args query text.`);
+  private static getAnswerDialog(min: GBMinInstance, service: KBService) {
+    return [
+      async step => {
+        const user = await min.userProfile.get(step.context, {});
+        let text = step.options.query;
+        if (!text) {
+          throw new Error(`/answer being called with no args query text.`);
+        }
+        const locale = step.context.activity.locale;
+        // Stops any content on projector.
+        await min.conversationalService.sendEvent(step, 'stop', undefined);
+        // Handle extra text from FAQ.
+        if (step.options && step.options.query) {
+          text = step.options.query;
+        } else if (step.options && step.options.fromFaq) {
+          await step.context.sendActivity(Messages[locale].going_answer);
+        }
+        // Spells check the input text before sending Search or NLP.
+        if (min.instance.spellcheckerKey !== undefined) {
+          const data = await AzureText.getSpelledText(min.instance.spellcheckerKey, text);
+          if (data !== text) {
+            GBLog.info(`Spelling corrected: ${data}`);
+            text = data;
           }
+        }
 
-          const locale = step.context.activity.locale;
+        // Searches KB for the first time.
+        user.lastQuestion = text;
+        await min.userProfile.set(step.context, user);
+        const resultsA = await service.ask(min.instance, text, min.instance.searchScore, user.subjects);
 
-          // Stops any content on projector.
-
-          await min.conversationalService.sendEvent(step, 'stop', null);
-
-          // Handle extra text from FAQ.
-
-          if (step.options && step.options['query']) {
-            text = step.options['query'];
-          } else if (step.options && step.options['fromFaq']) {
-            await step.context.sendActivity(Messages[locale].going_answer);
-          }
-
-          // Spells check the input text before sending Search or NLP.
-
-          if (min.instance.spellcheckerKey) {
-            const data = await AzureText.getSpelledText(min.instance.spellcheckerKey, text);
-
-            if (data != text) {
-              logger.info(`Spelling corrected: ${data}`);
-              text = data;
-            }
-          }
-
-          // Searches KB for the first time.
-
-          user.lastQuestion = text;
+        // If there is some result, answer immediately.
+        if (resultsA !== undefined && resultsA.answer !== undefined) {
+          // Saves some context info.
+          user.isAsking = false;
+          user.lastQuestionId = resultsA.questionId;
           await min.userProfile.set(step.context, user);
-          const resultsA = await service.ask(min.instance, text, min.instance.searchScore, user.subjects);
 
+          // Sends the answer to all outputs, including projector.
+          await service.sendAnswer(min.conversationalService, step, resultsA.answer);
+
+          // Goes to ask loop, again.
+          return await step.replaceDialog('/ask', { isReturning: true });
+        } else {
+          // Second time running Search, now with no filter.
+          const resultsB = await service.ask(min.instance, text, min.instance.searchScore, undefined);
           // If there is some result, answer immediately.
 
-          if (resultsA && resultsA.answer) {
+          if (resultsB !== undefined && resultsB.answer !== undefined) {
             // Saves some context info.
-
-            user.isAsking = false;
-            user.lastQuestionId = resultsA.questionId;
-            await min.userProfile.set(step.context, user);
-
+            const user2 = await min.userProfile.get(step.context, {});
+            user2.isAsking = false;
+            user2.lastQuestionId = resultsB.questionId;
+            await min.userProfile.set(step.context, user2);
+            // Informs user that a broader search will be used.
+            if (user2.subjects.length > 0) {
+              await step.context.sendActivity(Messages[locale].wider_answer);
+            }
             // Sends the answer to all outputs, including projector.
-
-            await service.sendAnswer(min.conversationalService, step, resultsA.answer);
-
-            // Goes to ask loop, again.
+            await service.sendAnswer(min.conversationalService, step, resultsB.answer);
 
             return await step.replaceDialog('/ask', { isReturning: true });
           } else {
-            // Second time running Search, now with no filter.
+            if (!(await min.conversationalService.routeNLP(step, min, text))) {
+              await step.context.sendActivity(Messages[locale].did_not_find);
 
-            const resultsB = await service.ask(min.instance, text, min.instance.searchScore, null);
-
-            // If there is some result, answer immediately.
-
-            if (resultsB && resultsB.answer) {
-              // Saves some context info.
-
-              const user = await min.userProfile.get(step.context, {});
-              user.isAsking = false;
-              user.lastQuestionId = resultsB.questionId;
-              await min.userProfile.set(step.context, user);
-
-              // Informs user that a broader search will be used.
-
-              if (user.subjects.length > 0) {
-                const subjectText = `${KBService.getSubjectItemsSeparatedBySpaces(user.subjects)}`;
-                await step.context.sendActivity(Messages[locale].wider_answer);
-              }
-
-              // Sends the answer to all outputs, including projector.
-
-              await service.sendAnswer(min.conversationalService, step, resultsB.answer);
               return await step.replaceDialog('/ask', { isReturning: true });
-            } else {
-              if (!(await min.conversationalService.routeNLP(step, min, text))) {
-                await step.context.sendActivity(Messages[locale].did_not_find);
-                return await step.replaceDialog('/ask', { isReturning: true });
-              }
             }
           }
         }
-      ])
-    );
+      }
+    ];
+  }
 
-    min.dialogs.add(
-      new WaterfallDialog('/ask', [
-        async step => {
-          const locale = step.context.activity.locale;
-          const user = await min.userProfile.get(step.context, {});
-          user.isAsking = true;
-          if (!user.subjects) {
-            user.subjects = [];
-          }
-          let text;
-
-          // Three forms of asking.
-
-          if (step.options && step.options['firstTime']) {
-            text = Messages[locale].ask_first_time;
-          } else if (step.options && step.options['isReturning']) {
-            text = Messages[locale].anything_else;
-          } else if (user.subjects.length > 0) {
-            text = Messages[locale].which_question;
-          } else {
-            throw new Error('Invalid use of /ask');
-          }
-
-          if (text.length > 0) {
-            return await step.prompt('textPrompt', text);
-          }
-          return await step.next();
-        },
-        async step => {
-          if (step.result) {
-            return await step.replaceDialog('/answer', { query: step.result });
-          } else {
-            return await step.next();
-          }
+  private static getAnswerEventDialog(service: KBService, min: GBMinInstance) {
+    return [
+      async step => {
+        const data = step.options as AskDialogArgs;
+        if (data !== undefined && data.questionId !== undefined) {
+          const question = await service.getQuestionById(min.instance.instanceId, data.questionId);
+          const answer = await service.getAnswerById(min.instance.instanceId, question.answerId);
+          // Sends the answer to all outputs, including projector.
+          await service.sendAnswer(min.conversationalService, step, answer);
+          await step.replaceDialog('/ask', { isReturning: true });
         }
-      ])
-    );
+
+        return await step.next();
+      }
+    ];
   }
 }
